Refactor some unit tests that had new class instantiations in the PHPUnit data provider.
On April 2, 2017, Travis CI began to fail on certain unit tests even though no code changes occured in those locations. It could not be reproduced locally, but was consistently failing in Travis CI. It was failing on unit tests that instantiated new objects in the PHPUnit data provider. The workaround was to move the object instantiations into the test method. It is unclear why this error started happening in Travis CI. As a future todo it would be nice to revert these changes as the test data providers are clearer with the objects.